[EPT-1709_MAKE_PAGE_TABLE_TIDIER] add hover options
